Fix error when downloading only covers and reviewing downloaded metadata. Fixes #1176253 (Bulk Download Covers Only - 0.9.29)

+++ src/calibre/gui2/actions/edit_metadata.py
@@ -240,9 +240,10 @@ class EditMetadataAction(InterfaceAction):
                         opf, cov = id_map[book_id]
                         cfile = mi.cover
                         mi.cover, mi.cover_data = None, (None, None)
-                        with open(opf, 'wb') as f:
-                            f.write(metadata_to_opf(mi))
-                        if cfile:
+                        if opf is not None:
+                            with open(opf, 'wb') as f:
+                                f.write(metadata_to_opf(mi))
+                        if cfile and cov:
                             shutil.copyfile(cfile, cov)
                             os.remove(cfile)
                     nid_map[book_id] = id_map[book_id]
